pyproject.toml
[tool.poetry]
name = "fintech-salon"
version = "0.1.0"
description = ""
authors = ["TonyLin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic = "^2.5.2"
openai = "^1.3.7"
bs4 = "^0.0.1"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
